history: Added history and refresh save
The file history.js keeps track of what is in the text box and also
listens for a click.  The rest was added for navigation once the website
is deployed.  It saves the last page loaded (window.onload).

Next to come: cross-domain XMLHttpRequest().  If I can figure out the
authentication methods required for the new API's.