wiki: search using AJAX and JSONP is up
The search string provided in the input box will query wikipedia
and send back 10 results. The query faced auth issues so the
data format was switched to JSONP; only GET requests are performed
on this site.

If deployed (for professional use): I would first implement a way
to get the next set of results and then make the snippets give a
better intro string.  Among other various search options.

Next: Will need to deploy the page on a web server using HTTPS
protocol to get any further with debugging the Twitter widget.