heroku: Deploy app to test twitter.js
Here I have implemented changes to the structure of the code in
order to deploy it on Heroku as a Node.js application.  This is the
only way I can test the twitter.js app seeing as how the Twitter
API requires that the origin uses HTTPS protocol.

Wish me luck..